app/requirements.txt
Django==3.1
psycopg2-binary==2.9.1
djangorestframework==3.11.0
Pillow==9.2.0
pytz==2019.3
sqlparse==0.3.1
drf-yasg==1.20.0